Moved -class to separate module
