+++ Source/GbEnum/__init__.py
@@ -1,0 +1,8 @@
+# -*- encoding: utf-8 -*-
+#
+# Authors: Thomas Li Fredriksen
+# License: See LICENSE file
+# GitHub: https://github.com/thomafred/PyBoy
+#
+
+from AutoEnum import AutoEnum
