+++ Source/WindowEvent.py
@@ -5,14 +5,7 @@
 # GitHub: https://github.com/Baekalfen/PyBoy
 #
 
-from enum import Enum
-
-class AutoEnum(Enum):
-    def __new__(cls):
-        value = len(cls.__members__) + 1
-        obj = object.__new__(cls)
-        obj._value_ = value
-        return obj
+from GbEnum import AutoEnum
 
 class WindowEvent(AutoEnum):
